feat: streamline component imports by consolidating them into a single index file
