+++ client/src/components/index.ts
@@ -1,0 +1,10 @@
+import ChannelInfo from './ChannelInfo.vue'
+import ChannelsList from './ChannelsList.vue'
+import CommandLine from './CommandLine.vue'
+import CommandsHelp from './CommandsHelp.vue'
+import DialogWrapper from './DialogWrapper.vue'
+import NewChannel from './NewChannel.vue'
+import UserMessage from './UserMessage.vue'
+import UserStatus from './UserStatus.vue'
+
+export { ChannelInfo, ChannelsList, CommandLine, CommandsHelp, DialogWrapper, NewChannel, UserMessage, UserStatus }
